+++ meinberlin/apps/account/views.py
@@ -1,3 +1,5 @@
+import json
+
 from django.contrib.messages.views import SuccessMessageMixin
 from django.shortcuts import get_object_or_404
 from django.urls import reverse
@@ -8,7 +10,10 @@ from rest_framework.renderers import JSONRenderer
 from rules.contrib.views import LoginRequiredMixin
 
 from adhocracy4.actions.models import Action
+from adhocracy4.projects.models import Project
+from meinberlin.apps.contrib.enums import TopicEnum
 from meinberlin.apps.notifications.serializers import NotificationSettingsSerializer
+from meinberlin.apps.plans.models import Topic
 from meinberlin.apps.users.models import User
 
 from . import forms
@@ -67,4 +72,37 @@ class NotificationSettingsView(LoginRequiredMixin, generic.TemplateView):
             "notification-settings-detail",
             kwargs={"pk": user.notification_settings.id},
         )
+        return context
+
+
+class FollowedProjectsListView(LoginRequiredMixin, generic.ListView):
+    template_name = "meinberlin_kiezradar/followed_projects_list.html"
+    context_object_name = "followed_projects"
+
+    def get_topics(self):
+
+        topics = [
+            {
+                "id": topic.id,
+                "code": topic.code,
+                "name": str(TopicEnum(topic.code).label),
+            }
+            for topic in Topic.objects.all()
+        ]
+        return json.dumps(topics)
+
+    def get_queryset(self):
+        user = self.request.user
+        # Get all projects that the user follows with enabled=True
+        return (
+            Project.objects.filter(follow__creator=user, follow__enabled=True)
+            .prefetch_related("topics")
+            .distinct()
+            .order_by("-created")
+        )
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["projects_api_url"] = reverse("projects-list")
+        context["topic_choices"] = self.get_topics()
         return context